added perform action but doesn't work
